refactor(pipeline): decompose BasePipeline into focused modules
- Extract lock management to lock_manager.py (PipelineLockManager)
- Extract checkpoint handling to checkpoint_manager.py (PipelineCheckpointManager)
- Extract record processing to record_processor.py (PipelineRecordProcessor)
- Move ChEMBLActivityPipelineFactory to infrastructure/factories/
- Reduce base.py from 315 LOC to 99 LOC
- Remove ignore_imports from .importlinter (clean architecture)

Addresses R3 (decomposition) and R4 (import-linter cleanup).
