+++ src/bioetl/application/pipeline/__init__.py
@@ -1,1 +1,19 @@
 """Pipeline components and base classes."""
+
+from bioetl.application.pipeline.base import (
+    BasePipeline,
+    PipelineShutdownError,
+    run_pipeline_flow,
+)
+from bioetl.application.pipeline.checkpoint_manager import PipelineCheckpointManager
+from bioetl.application.pipeline.lock_manager import PipelineLockManager
+from bioetl.application.pipeline.record_processor import PipelineRecordProcessor
+
+__all__ = [
+    "BasePipeline",
+    "PipelineShutdownError",
+    "run_pipeline_flow",
+    "PipelineCheckpointManager",
+    "PipelineLockManager",
+    "PipelineRecordProcessor",
+]
