+++ src/bioetl/infrastructure/factories/__init__.py
@@ -1,0 +1,1 @@
+"""Pipeline factories for dependency injection."""
